packages/auth/requirements.txt
flask
flask-cors
flask-jwt-extended
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.0